+++ src/data/navigation/sections/admin-ui-sdk.js
@@ -21,19 +21,19 @@ module.exports = [
         },
         {
           title: "order",
-          path: "/admin-ui-sdk/extension-points/order.md",
+          path: "/admin-ui-sdk/extension-points/order/index.md",
           pages: [
             {
-              title: "grid column",
-              path: "/admin-ui-sdk/extension-points/grid-columns.md"
+              title: "grid columns",
+              path: "/admin-ui-sdk/extension-points/order/grid-columns.md"
             },
             {
               title: "mass action",
-              path: "admin-ui-sdk/extension-points/mass-action.md"
+              path: "/admin-ui-sdk/extension-points/order/mass-action.md"
             },
             {
               title: "view button",
-              path: "admin-ui-sdk/extension-points/view-button.md"
+              path: "/admin-ui-sdk/extension-points/order/view-button.md"
             }
           ]
         },
@@ -43,7 +43,17 @@ module.exports = [
         },
         {
           title: "product",
-          path: "/admin-ui-sdk/extension-points/product.md"
+          path: "/admin-ui-sdk/extension-points/product/index.md",
+          pages: [
+            {
+              title: "grid columns",
+              path: "/admin-ui-sdk/extension-points/product/grid-columns.md"
+            },
+            {
+              title: "mass action",
+              path: "/admin-ui-sdk/extension-points/product/mass-action.md"
+            }
+          ]
         }
 
       ]